Fixes
Path.IsRootedPath was not the correct way to attempt to establish abs
paths. Also Path.PathSeparator was not correct, what we're after is the
directory path separator, aka forward or back slash.

Update core project to include command line parser as a dep. We may want
to enable dynamic command line requirements in the future, so it's added
to the core lib rather than buildbot.